DS-12 Move some definitions from DragNDriopBehavior to DraggableShapeBehavior
